Describe delete interest tests with delete modal component
instead of EditInterest to avoid a confusing error with network requests
